ewawadwa

+++ database/migrations/2025_04_15_123947_add_column_to_users_table.php
@@ -14,10 +14,12 @@ return new class extends Migration
         
     }
 
-    public function down(): void
+    public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('donasi_count');
-        });
+        if (Schema::hasTable('users') && Schema::hasColumn('users', 'donasi_count')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('donasi_count');
+            });
+        }
     }
 };
